+++ src/apyfloat_wrapper.cc
@@ -593,5 +593,83 @@ void bind_float(nb::module_& m)
             -------
             :class:`APyFloat`
 
-            )pbdoc");
+            )pbdoc")
+        .def_static(
+            "inf",
+            &APyFloat::inf,
+            nb::arg("exp_bits"),
+            nb::arg("man_bits"),
+            nb::arg("bias") = nb::none(),
+            R"pbdoc(
+            Create an :class:`APyFloat` object that is initialized to positive infinity.
+
+            versionadded:: 0.4
+
+            Parameters
+            ----------
+            exp_bits : :class:`int`
+                Number of exponent bits.
+            man_bits : :class:`int`
+                Number of mantissa bits.
+            bias : :class:`int`, optional
+                Exponent bias. If not provided, *bias* is ``2**exp_bits - 1``.
+
+            Examples
+            --------
+
+            >>> import apytypes as apy
+
+            `a`, initialized to positive infinity.
+
+            >>> a = apy.APyFloat.inf(exp_bits=10, man_bits=15)
+
+            Returns
+            -------
+            :class:`APyFloat`
+
+            See also
+            --------
+            nan
+
+            )pbdoc"
+        )
+        .def_static(
+            "nan",
+            &APyFloat::nan,
+            nb::arg("exp_bits"),
+            nb::arg("man_bits"),
+            nb::arg("bias") = nb::none(),
+            R"pbdoc(
+            Create an :class:`APyFloat` object that is initialized to NaN.
+
+            versionadded:: 0.4
+
+            Parameters
+            ----------
+            exp_bits : :class:`int`
+                Number of exponent bits.
+            man_bits : :class:`int`
+                Number of mantissa bits.
+            bias : :class:`int`, optional
+                Exponent bias. If not provided, *bias* is ``2**exp_bits - 1``.
+
+            Examples
+            --------
+
+            >>> import apytypes as apy
+
+            `a`, initialized to NaN.
+
+            >>> a = apy.APyFloat.nan(exp_bits=10, man_bits=15)
+
+            Returns
+            -------
+            :class:`APyFloat`
+
+            See also
+            --------
+            inf
+
+            )pbdoc"
+        );
 }