+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `inf` and `nan` functions to create floating-point objects initialized
+  to infinity or NaN.
+
 ### Fixed
 
 ### Changed